Updated to Version 2.1
Added some switches.
DisabledOnly, gets only disabled users. Makes it faster.
LicenseOnly, gets only the license information for the user. Makes it a lot faster.
Set the AD switch to be an Include rather than exclude, so it's -IncludeAD if you want to add more details and you have access to the internal AD environment.